added documentation,refactor

+++ src/main/java/com/myproject/dao/DriverDao.java
@@ -4,6 +4,6 @@ import com.myproject.exception.DaoException;
 
 public interface DriverDao {
 
-    boolean updateDriverPrice(int price) throws DaoException;
+    boolean updateDriverPrice(double price) throws DaoException;
     double getDriverPrice() throws DaoException;
 }
